remove deprecated pages
