--- packages/web/src/pages/[network]/create/github/addresses.json
@@ -1,8 +1,0 @@
-{
-  "ethereum": "0x34A7AdC94C4D41C3e3469F98033B372cB2fAf318",
-  "ropsten": "0x1A45BA7BB4a5A4a01Eb5330C9F2fd8af47C16eBe",
-  "polygon": "0xa7d0A25Cb2378c935FF94863C8614367b9e85Ed8",
-  "arbitrum-one": "0x84b9e407e2Ee76A641b45d007bBFd9e60a13FF7d",
-  "polygon-mumbai": "0x9edc758B0477559AB6556BCA9971B4DD892d8E28",
-  "arbitrum-rinkeby": "0x82045e7aD20dca31915896ED246bB56a3B5956d8"
-}
